requirements.txt
selenium==4.10.0
pytest==7.4.0
pytest-html==3.2.0
allure-pytest==2.12.0
webdriver-manager==4.0.2
kubernetes==31.0.0